Implementing bounding volume hierarchy nodes
